renameContainerToFAB
